using Int coins

+++ src/main/scala/data/InventoryItem.scala
@@ -3,4 +3,4 @@ package data
 /**
   * Created by adam on 8/24/16.
   */
-case class InventoryItem(product : Product, price : BigDecimal, quantity : Int)
+case class InventoryItem(product : Product, price : Int, quantity : Int)
